refactor: move all production code to use new function
